refactor: moved several js files to src/
